backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
httpx==0.28.1
supabase==2.18.1
cryptography==45.0.7
pydantic-settings==2.10.1
pytest==9.0.2
